requirements-dashboard.txt
# Streamlit Dashboard Requirements (Python 3.8 / master environment)
# Install with: pip install -r requirements-dashboard.txt

# Core Dashboard
streamlit==1.40.1
altair>=5.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.20.0

# NLP
nltk>=3.8.1

# OpenAI for RAG+LLM (Legacy 0.28.x compatibility)
openai==0.28.0

# Optional: for TF-IDF retrieval method
scikit-learn>=1.3.0
